+++ Migrations/Code/ClassAliasMap.php
@@ -1,6 +1,6 @@
 <?php
 
 return [
-    'Tx_Phpunit_Interface_AccessibleObject' => \OliverKlee\Phpunit\Interfaces\AccessibleObject::class,
+    'Tx_Phpunit_Interface_AccessibleObject' => \OliverKlee\PhpUnit\Interfaces\AccessibleObject::class,
     'Tx_Phpunit_TestCase' => \OliverKlee\PhpUnit\TestCase::class,
 ];