+++ Tests/Unit/BackEnd/TestListenerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OliverKlee\Phpunit\Tests\Unit\BackEnd;
+namespace OliverKlee\PhpUnit\Tests\Unit\BackEnd;
 
 use OliverKlee\PhpUnit\TestCase;
 use SebastianBergmann\Comparator\ComparisonFailure;