+++ TestExtensions/bbb/ext_emconf.php
@@ -3,7 +3,7 @@
 $EM_CONF[$_EXTKEY] = [
     'title' => 'BBB',
     'description' => 'A test extension used for running the phpunit unit tests.',
-    'version' => '5.3.5',
+    'version' => '5.7.27',
     'category' => 'example',
     'constraints' => [
         'depends' => [