+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Run Travis CI with highest and lowest dependencies (#107)
 - Add option to pass CLI options to PHPUnit (#102)
 - Add TestCase::getProtectedProperty (#91)
 - Auto-release to the TER (#67)
@@ -35,7 +36,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Remove the obsolete ext_autoload.php (#56)
 
 ### Fixed
-- Fix code inspection warnings (#105)
+- Fix code inspection warnings (#106)
 - Fix the casing of the RunTestsCommand namespace (#101)
 - Stop using the Core base exception class (#100)
 - Fix the autoloading in RunTestsCommand (#95)